*: move etcdserver/idutil -> pkg/idutil
